adding more lexical environment records
